Renamed connect.resource to connect.resources and created connect.exceptions (although connect.models.exception still contains the old ones).
