+++ connect/__init__.py
@@ -5,16 +5,17 @@ This file is part of the Ingram Micro Cloud Blue Connect SDK.
 Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 """
 
-from .resource import FulfillmentAutomation, TierConfigAutomation
+from .resources import FulfillmentAutomation, TierConfigAutomation
 
 name = 'connect'
 
 __all__ = [
     'config',
+    'exceptions',
     'logger',
     'migration_handler',
     'models',
-    'resource',
+    'resources',
 
     # TODO: Provided for backwards compatibility
     'FulfillmentAutomation',
