+++ connect/models/__init__.py
@@ -13,9 +13,6 @@ from .connection import Connection, ConnectionSchema
 from .contact import Contact, ContactInfo, ContactInfoSchema, ContactSchema, \
     PhoneNumber, PhoneNumberSchema
 from .event import EventInfo, EventInfoSchema, Events, EventsSchema
-from .exception import AcceptUsageFile, CloseUsageFile, DeleteUsageFile, FailRequest, \
-    FileCreationError, FileRetrievalError, FulfillmentFail, FulfillmentInquire, InquireRequest, \
-    Message, RejectUsageFile, ServerError, SkipRequest, Skip, SubmitUsageFile, UsageFileAction
 from .fulfillment import Fulfillment, FulfillmentSchema
 from .hub import Hub, HubInstance, HubInstanceSchema, Hubs, HubSchema, HubsSchema, HubStats, \
     HubStatsSchema
@@ -34,7 +31,6 @@ from .usage import File, FileSchema, FileUsageRecord, FileUsageRecordSchema, \
     Listing, ListingSchema, Records, RecordsSchema
 
 __all__ = [
-    'AcceptUsageFile',
     'Account',
     'AccountSchema',
     'Activation',
@@ -49,7 +45,6 @@ __all__ = [
     'AssetSchema',
     'BaseModel',
     'BaseSchema',
-    'CloseUsageFile',
     'Company',
     'CompanySchema',
     'Connection',
@@ -64,7 +59,6 @@ __all__ = [
     'ContractSchema',
     'CustomerUiSettings',
     'CustomerUiSettingsSchema',
-    'DeleteUsageFile',
     'Document',
     'DocumentSchema',
     'DownloadLink',
@@ -73,16 +67,11 @@ __all__ = [
     'EventInfoSchema',
     'Events',
     'EventsSchema',
-    'FailRequest',
     'File',
-    'FileCreationError',
-    'FileRetrievalError',
     'FileSchema',
     'FileUsageRecord',
     'FileUsageRecordSchema',
     'Fulfillment',
-    'FulfillmentFail',
-    'FulfillmentInquire',
     'FulfillmentSchema',
     'Hub',
     'HubInstance',
@@ -92,14 +81,12 @@ __all__ = [
     'HubsSchema',
     'HubStats',
     'HubStatsSchema',
-    'InquireRequest',
     'Item',
     'ItemSchema',
     'Listing',
     'ListingSchema',
     'Marketplace',
     'MarketplaceSchema',
-    'Message',
     'Param',
     'ParamSchema',
     'PhoneNumber',
@@ -110,15 +97,10 @@ __all__ = [
     'ProductSchema',
     'Records',
     'RecordsSchema',
-    'RejectUsageFile',
     'Renewal',
     'RenewalSchema',
-    'ServerError',
     'ServerErrorResponse',
     'ServerErrorResponseSchema',
-    'Skip',
-    'SkipRequest',
-    'SubmitUsageFile',
     'Template',
     'TemplateSchema',
     'Tier',
@@ -129,7 +111,6 @@ __all__ = [
     'Tiers',
     'TierSchema',
     'TiersSchema',
-    'UsageFileAction',
     'ValueChoice',
     'ValueChoiceSchema',
 ]
