+++ connect/resources/automation_engine.py
