+++ connect/resources/base.py
@@ -11,8 +11,9 @@ import requests
 from requests import compat
 
 from connect.config import Config
+from connect.exceptions import ServerError
 from connect.logger import function_log, logger
-from connect.models import BaseSchema, ServerError, ServerErrorResponseSchema
+from connect.models import BaseSchema, ServerErrorResponseSchema
 
 
 class ApiClient(object):
