+++ connect/resources/tier_config_automation.py
@@ -8,9 +8,10 @@ from abc import ABCMeta
 
 from typing import List
 
+from connect.exceptions import FailRequest, InquireRequest, SkipRequest
 from connect.logger import logger, function_log
-from connect.models import ActivationTemplateResponse, ActivationTileResponse, Param, FailRequest,\
-    InquireRequest, SkipRequest, TierConfigRequest, TierConfigRequestSchema
+from connect.models import ActivationTemplateResponse, ActivationTileResponse, Param, \
+    TierConfigRequest, TierConfigRequestSchema
 from .automation_engine import AutomationEngine
