+++ connect/resources/usage_file_automation.py
@@ -6,9 +6,9 @@ Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 """
 from abc import ABCMeta
 
+from connect.exceptions import SkipRequest, UsageFileAction
 from connect.logger import logger
 from connect.models.base import BaseModel
-from connect.models.exception import UsageFileAction, SkipRequest
 from connect.models.usage import FileSchema, File
 from .automation_engine import AutomationEngine
